+++ CHANGELOG.md
@@ -2,7 +2,25 @@
 
 ## develop
 
-# v0.1.1
+## v0.2.0
+
+Released Wednesday, 29th January 2020.
+
+### Backwards-Compatibility Break
+
+* `Uuid` is now a branded string. No longer a class / value object.
+* `mustBeUuid()` is now `mustBeUuidData()`
+* `mustBeUuidWithOnError()` has been removed; use `mustBeUuidData()` instead
+
+### New
+
+* Added automatic to-string conversion of Uuids.
+
+### Refactor
+
+* Move conversion to/from unformatted strings into separate functions
+
+## v0.1.1
 
 Released Thursday, 2nd January 2020.
 
@@ -10,7 +28,7 @@ Released Thursday, 2nd January 2020.
 
 * Solved the problem of missing Typescript declaration files from the published package.
 
-# v0.1.0
+## v0.1.0
 
 Released Thursday, 2nd January 2020.